backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
asyncpg
passlib[bcrypt]
python-jose[cryptography]
